Moved werkzueg functions from db into admin.
Patches circular import error that started occuring after I seperated the database code into a seperate module. Fixed it by moving the check_password_hash from within the database to within the admin blueprint. This function is used for logging users in. Now when a new admin is added to the database I have to manually create the password hash. I plan to create a 'register' route to make this simpler. I'm not in any hurry because I am the only admin. Updated tests.